Finish getting completedAt timestamp working

+++ server/public/scripts/client.js
@@ -82,22 +82,25 @@ function toggleStatus(event, todoId) {
 // Make DELETE HTTP request to delete todo, refresh from db and re-render
 function deleteTodo(todoId) {
 
-  // Send Axios HTTP DELETE request to server
-  axios({
-    method: "DELETE",
-    url: "/todos",
-    data: { id: todoId }
-  })
-    .then((response) => {
-      // DELETE request and response successful so get updated todos and re-render to DOM
-      // console.log("/todos DELETE request sent - response from server:", response.data);
-      getTodos();
-  })
-    .catch((error) => {
-      // Log error
-      console.log("/todos DELETE request encountered an error:", error);
-  });
-
+  // Delete todo if user confirms
+  const confirmDelete = confirm(`‼️ Are you sure you want to delete this to-do?`);
+  if (confirmDelete) {
+    // Send Axios HTTP DELETE request to server
+    axios({
+      method: "DELETE",
+      url: "/todos",
+      data: { id: todoId }
+    })
+      .then((response) => {
+        // DELETE request and response successful so get updated todos and re-render to DOM
+        // console.log("/todos DELETE request sent - response from server:", response.data);
+        getTodos();
+    })
+      .catch((error) => {
+        // Log error
+        console.log("/todos DELETE request encountered an error:", error);
+    });
+  }
 }
 
 // Render todos
@@ -109,18 +112,31 @@ function renderTodos(todos) {
   // Loop through todos received from database and add to DOM
   for (const todo of todos) {
 
+    // Format the date
+    const timestampz = todo.completedAt;
+    const date = new Date(timestampz);
+    const args = {
+      year: 'numeric',
+      month: 'short',
+      day: 'numeric',
+      hour: 'numeric',
+      minute: 'numeric'
+    };
+    const timeStamp = new Intl.DateTimeFormat('en-US', args).format(date);
+
     // Handle look of todo based on completed status
     const isChecked = todo.isComplete ? true : false;
     const liClassList = todo.isComplete ? ` bg-transparent fst-italic text-secondary completed`: ` bg-dark-subtle`;
     const btnClassList = todo.isComplete ? ` btn-outline-success` : ` btn-outline-secondary`;
     const checkmarkHtml = todo.isComplete ? `<svg class="text-success" fill="currentColor" xmlns="http://www.w3.org/2000/svg" xml:space="preserve" width="800" height="800" viewBox="0 0 17.837 17.837"><path d="M16.145 2.571a.7.7 0 0 0-.99 0L6.92 10.804l-4.241-4.27a.698.698 0 0 0-.989 0L.204 8.019a.703.703 0 0 0 0 .99l6.217 6.258a.704.704 0 0 0 .99 0L17.63 5.047a.7.7 0 0 0 0-.994l-1.485-1.482z"/></svg>` : ``;
     const badgeHtml = todo.isComplete ? `<span class="badge rounded-pill bg-transparent text-success border border-success fst-normal fw-light">Completed</span>` : ``;
+    const completionTimeHtml = todo.isComplete ? `<small class="me-4 fst-italic text-dark-subtle lh-1">Completed ${timeStamp}</small>` : ``;
 
     // Append retrieved todos to list
     todoListUl.innerHTML += `
       <li class="list-group-item${liClassList}" data-testid="toDoItem">
         <div class="row g-3">
-          <div class="col-8 d-flex align-items-center">
+          <div class="col-7 d-flex align-items-center">
             <button id="completeBtn" onclick="toggleStatus(event, ${todo.id})" class="btn${btnClassList}" data-completed="${isChecked}" data-testid="completeButton">
               ${checkmarkHtml}
             </button>
@@ -129,7 +145,8 @@ function renderTodos(todos) {
               <div class="ms-3">${badgeHtml}</div>
             </div>
           </div>
-          <div class="col-4 d-flex justify-content-end">
+          <div class="col-5 d-flex align-items-center justify-content-end">
+            ${completionTimeHtml}
             <button onclick="deleteTodo(${todo.id});" class="btn btn-sm btn-outline-danger" data-testid="deleteButton">Delete</button>
           </div>
         </div>
